Update sparse indexing code in temoa_initialize.py
Code in temoa_initialize.py was updated to eliminate a series of function calls related to sparse matrix indexing. These functions were simply returning a set of indices from a pre-defined dictionary. Now, the function calls are directly replaced by reading the appropriate dictionaries. Replacing function calls with dictionaries also needs to be done in temoa_rules.py, which will be handled in the next commit.