Fix calculation of V_FlowIn in pformat_results.py
The model allows for curtailment of technologies defined in the tech_curtailment set; however, this curtailed energy was not factored into the calculation of V_FlowIn. Rather than explicitly add curtailment to the CommodityBalance constraints, we instead calculate V_FlowIn (accounting for curtailed energy) ex post in pformat_results.py. There is one important caveat to our approach: members of the tech_curtailment set should not have an upstream cost, as the curtailed energy is not explicitly accounted for when doing the commodity balance during the optimization. As long as the input commodity is free, as in the case of wind and solar, this imbalance is fine.