Update storage to account for time slice length
This commit corrects an issue in the storage algorithm, which needs to account for varying, user-defined time slice lengths. This issue is fixed by considering the number of days within each season. Also, the storage duration should now be specified in units of hours rather than fraction of a year, as it was defined previously.