Optimize initial storage charge levels
Instead of specifying arbitrary values for initial storage charge levels, we now optimize them. In order to do this, StorageInit[t,v] is defined as a decision variable that determines the initial storage level. There is no need to add any new constraints because StorageEnergyUpperBound_Constraint requires StorageInit[t,v] to be smaller than the 'energy_capacity'. Consistent with the previous format, we still require all the storage technologies to have equal initial and final storage levels. However, to preserve the ability of modelers to force a specific initial charge level as a fraction of ‘energy_capacity’, a new parameter is defined: StorageInitFrac[t,v]. If a value is specified in the input database, this parameter overrides the optimization using the newly created StorageInit_Constraint.