Add maximum resource constraint
A new maximum resource constraint was added to the model.
This constraint is similar to the max activity constraint,
but is only indexed by region and technology, and not by
time period. This accounts for cases where the total resource
capacity of a particular technology/commodity combination over
the entire modeling horizon is constrained. Note that this will
work correctly for technologies that have a single output commodity.

The temoa_schema, temoa_test_system and temoa_utopia databases
were updated to include the MaxResource table.