Complete first step of adding region index to Graphviz module
We're in the process of adding the region index to the Graphviz code so that users can produce input and output network diagrams by region. This first commit adds the region flag and allows users to generate the full input diagram accounting for region. Additional commits will add consideration of the region to technology and commodity input sub-diagrams, as well as all the output network diagrams.