Remove commodity_SNG as it's not needed anymore
commodity_SNG represented the union of physical and emissions commodities, which in turn served as the commodity input to the Efficiency table. This approach allowed for the CO2 emissions commodity to serve as an input to the production of synthetic natural gas (SNG). This commodity type is no longer needed and instead was replaced with LinkedEmissionsTech_Constraint.