Fix primary/linked techs in myopic solution
The solvers sometimes report small values for linked technologies
but not the primary technology, or vice versa. This causes issues
in future period solves where the capacity for one of the technologies
exists but not the other. A previous commit looks for such instances
and removes related data from the output tables. This commit adds to the
previous one where in some instances the primary tech exists but the linked
tech does not.